mmc: dw_mmc: Avoid cast mismatch build warning
This patch fixes below compile warning.

In file included from include/linux/irqflags.h:14:0,
   from ./arch/arm/include/asm/bitops.h:27,
   from include/linux/bitops.h:36,
   from include/linux/kernel.h:10,
   from include/linux/sched.h:17,
   from include/linux/blkdev.h:4,
   from drivers/mmc/host/dw_mmc.c:14:
drivers/mmc/host/dw_mmc.c: In function 'dw_mci_set_drto':
include/linux/typecheck.h:11:18: warning: comparison of distinct pointer types
lacks a cast [enabled by default]
  (void)(&__dummy == &__dummy2); \
                  ^
include/linux/spinlock.h:207:3: note: in expansion of macro 'typecheck'
  typecheck(unsigned long, flags); \
  ^
include/linux/spinlock.h:337:2: note: in expansion of macro 'raw_spin_lock_irqsave'
  raw_spin_lock_irqsave(spinlock_check(lock), flags); \
  ^
drivers/mmc/host/dw_mmc.c:1888:2: note: in expansion of macro 'spin_lock_irqsave'
  spin_lock_irqsave(&host->irq_lock, irqflags);
  ^

+++ drivers/mmc/host/dw_mmc.c
@@ -1873,7 +1873,7 @@ static void dw_mci_set_drto(struct dw_mci *host)
 	unsigned int drto_clks;
 	unsigned int drto_div;
 	unsigned int drto_ms;
-	unsigned int irqflags;
+	unsigned long irqflags;
 
 	drto_clks = mci_readl(host, TMOUT) >> 8;
 	drto_div = (mci_readl(host, CLKDIV) & 0xff) * 2;
